Split out DeveloperToolsScreen from GameSettingsScreen.cpp
