+++ CMakeLists.txt
@@ -1565,6 +1565,8 @@ list(APPEND NativeAppSource
 	UI/GameScreen.cpp
 	UI/GameSettingsScreen.h
 	UI/GameSettingsScreen.cpp
+	UI/DeveloperToolsScreen.h
+	UI/DeveloperToolsScreen.cpp
 	UI/DriverManagerScreen.h
 	UI/DriverManagerScreen.cpp
 	UI/GPUDriverTestScreen.h
