+++ UI/UI.vcxproj.filters
@@ -116,6 +116,9 @@
     <ClCompile Include="ImDebugger\ImConsole.cpp">
       <Filter>ImDebugger</Filter>
     </ClCompile>
+    <ClCompile Include="DeveloperToolsScreen.cpp">
+      <Filter>Screens</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -231,6 +234,9 @@
     </ClInclude>
     <ClInclude Include="ImDebugger\ImConsole.h">
       <Filter>ImDebugger</Filter>
+    </ClInclude>
+    <ClInclude Include="DeveloperToolsScreen.h">
+      <Filter>Screens</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
