+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -57,6 +57,7 @@
     <ClCompile Include="..\..\UI\ImDebugger\ImConsole.cpp">
       <Filter>ImDebugger</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\UI\DeveloperToolsScreen.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -115,6 +116,7 @@
     <ClInclude Include="..\..\UI\ImDebugger\ImConsole.h">
       <Filter>ImDebugger</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\UI\DeveloperToolsScreen.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="ImDebugger">
